consolidate vectorDb and agentExperiences
